requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# CLI (original)
click

# HTTP Client
httpx~=0.27.2

# Input handling
inputimeout==1.0.4

# JSON Schema
jsonschema==4.23.0
jsonschema-specifications==2023.12.1

# LangChain - LLM Integration
langchain==0.2.11
langchain-anthropic
langchain-huggingface
langchain-community==0.2.10
langchain-core==0.2.36
langchain-google-genai==1.0.10
langchain-ollama==0.1.3
langchain-openai==0.1.17
langchain-text-splitters==0.2.2
langsmith==0.1.93

# String matching
Levenshtein==0.25.1

# Logging
loguru==0.7.2

# OpenAI
openai==1.37.1

# PDF Processing
pdfminer.six==20221105
reportlab==4.2.2

# Environment
python-dotenv~=1.0.1

# YAML
PyYAML~=6.0.2

# Regex
regex==2024.7.24

# Pydantic for data validation
pydantic[email]>=2.0.0

# Testing (optional - for development)
pytest>=8.3.3
pytest-mock
pytest-cov

# Inquirer for CLI prompts
inquirer>=3.1.0
